requirements.txt
python-dotenv==1.0.1
requests==2.31.0
python-telegram-bot[webhooks]==22.7
black==24.4.2